+++ _test/test/goldens/generated_build_script.dart
@@ -75,7 +75,14 @@ final _builders = <_i1.BuilderApplication>[
       isOptional: true,
       hideOutput: true,
       appliesBuilders: ['build_modules:module_cleanup']),
-  _i1.apply('build_web_compilers:ddc', [_i7.ddcKernelBuilder, _i7.ddcBuilder],
+  _i1.apply(
+      'build_web_compilers:ddc',
+      [
+        _i7.ddcKernelBuilderUnsound,
+        _i7.ddcBuilderUnsound,
+        _i7.ddcKernelBuilderSound,
+        _i7.ddcBuilderSound
+      ],
       _i1.toAllPackages(),
       isOptional: true,
       hideOutput: true,
@@ -84,10 +91,12 @@ final _builders = <_i1.BuilderApplication>[
         'build_web_compilers:dart2js_modules',
         'build_web_compilers:dart_source_cleanup'
       ]),
-  _i1.apply('build_web_compilers:sdk_js_copy', [_i7.sdkJsCopyBuilder],
+  _i1.apply(
+      'build_web_compilers:sdk_js',
+      [_i7.sdkJsCompileUnsound, _i7.sdkJsCompileSound, _i7.sdkJsCopyRequirejs],
       _i1.toNoneByDefault(),
-      hideOutput: true,
-      appliesBuilders: ['build_web_compilers:sdk_js_cleanup']),
+      isOptional: true,
+      hideOutput: true),
   _i1.apply('build_web_compilers:entrypoint', [_i7.webEntrypointBuilder],
       _i1.toRoot(),
       hideOutput: true,
@@ -119,10 +128,6 @@ final _builders = <_i1.BuilderApplication>[
       defaultGenerateFor: const _i3.InputSet()),
   _i1.applyPostProcess(
       'build_web_compilers:dart_source_cleanup', _i7.dartSourceCleanup,
-      defaultReleaseOptions: _i8.BuilderOptions({'enabled': true}),
-      defaultGenerateFor: const _i3.InputSet()),
-  _i1.applyPostProcess(
-      'build_web_compilers:sdk_js_cleanup', _i7.sdkJsCleanupBuilder,
       defaultReleaseOptions: _i8.BuilderOptions({'enabled': true}),
       defaultGenerateFor: const _i3.InputSet()),
   _i1.applyPostProcess(